requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
ecdsa==0.18.0
fastapi==0.85.2
h11==0.14.0
idna==3.4
numpy==1.23.4
pandas==1.5.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
